algunos cambios

+++ src/Sistema.java
@@ -1,5 +1,6 @@
 import java.time.LocalDate;
 import java.time.chrono.ChronoLocalDate;
+import java.time.temporal.ChronoUnit;
 import java.util.LinkedList;
 import java.util.Scanner;
 
@@ -51,58 +52,91 @@ public class Sistema {
     public void Top_7_artistas() {
         MyList<Artista> artistaFechas = new MyLinkedListImpl<>();
         MyHeap<Artista, Artista> heapArtistas = new MyHeapIMPL<>();
-        System.out.println("------------------------------------------------------");
-        Scanner scanner1 = new Scanner(System.in);
-        System.out.println("Seleccione la fehca de inicio");
-        String fecha1 = scanner1.nextLine();
+        int cantcan = 0;
+        LocalDate localDate = null;
+        LocalDate localDate2 = null;
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
-        LocalDate localDate = LocalDate.parse(fecha1, formatter);
+        while (localDate == null) {
+            Scanner scanner1 = new Scanner(System.in);
+            System.out.println("Seleccione la fehca de inicio con el formato dd/mm/yyyy");
+            String fecha1 = scanner1.nextLine();
+            try {
+                localDate = LocalDate.parse(fecha1, formatter);
+            } catch (Exception e) {
+                System.out.println("La fecha no es compatible con el formato");
+            }
+
+        }
         ChronoLocalDate fecha1Formato = localDate;
-        Scanner scanner2 = new Scanner(System.in);
-        System.out.println("Seleccione la fehca de fin");
-        String fecha2 = scanner2.nextLine();
-        LocalDate localDate2 = LocalDate.parse(fecha2, formatter);
+        while (localDate2 == null) {
+            Scanner scanner2 = new Scanner(System.in);
+            System.out.println("Seleccione la fehca de fin con el formato dd/mm/yyyy");
+            String fecha2 = scanner2.nextLine();
+            try {
+                localDate2 = LocalDate.parse(fecha2, formatter);
+            } catch (Exception e) {
+                System.out.println("La fecha no es compatible con el formato");
+            }
+        }
         ChronoLocalDate fecha2Formato = localDate2;
         ChronoLocalDate tempfecha = fecha1Formato;
-        MyList<Top50> recorrer = arbolBusqueda.find(tempfecha).Values();
-        while (!tempfecha.equals(fecha2Formato)) {
-            for (int i = 0; i < recorrer.size(); i++) {
-                Top50 tempTop = recorrer.get(i);
-                MyHeap<Cancion, Integer> clonTop = tempTop.getTop().clonar();
-                for (int j = 0; j < 50; j++) {
-                    Cancion canciontemp = clonTop.delete();
-                    for (int a = 0; a < canciontemp.getArtist().size(); a++) {
-                        String nombreA = String.valueOf(canciontemp.getArtist().get(a));
-                        try {
-                            Artista tempArtista = procesador.hash().get(nombreA);
-                            if(!artistaFechas.contains(tempArtista)){
-                                artistaFechas.add(tempArtista);
-                            }
-                        }catch (NoEsta e){}
+        LocalDate localdatepri = LocalDate.parse("18/10/2023", formatter);
+        ChronoLocalDate primeraFecha = localdatepri;
+        LocalDate localdateult = LocalDate.parse("14/05/2024", formatter);
+        ChronoLocalDate ultimaFecha = localdateult;
+        if (fecha1Formato.isAfter(fecha2Formato)) {
+            System.out.println("No tiene sentido el orden");
+        }
+        if (fecha1Formato.isBefore(primeraFecha)) {
+            System.out.println("No tenemos registro de una fecha tan lejana");
+        }
+        if (fecha2Formato.isAfter(ultimaFecha)) {
+            System.out.println("No tenemos registro de una fecha tan cercana");
+        } else {
+            while (!tempfecha.equals(fecha2Formato)) {
+                if (arbolBusqueda.find(tempfecha) != null) {
+                    MyList<Top50> recorrer = arbolBusqueda.find(tempfecha).Values();
+                    for (int i = 0; i < recorrer.size(); i++) {
+                        Top50 tempTop = recorrer.get(i);
+                        MyHeap<Cancion, Integer> clonTop = tempTop.getTop().clonar();
+                        int tam = clonTop.size();
+                        for (int j = 0; j < tam; j++) {
+                            Cancion canciontemp = clonTop.delete();
+                            for (int a = 0; a < canciontemp.getArtist().size(); a++) {
+                                String nombreA = String.valueOf(canciontemp.getArtist().get(a));
+                                try {
+                                    Artista tempArtista = procesador.hash().get(nombreA);
+                                    if (!artistaFechas.contains(tempArtista)) {
+                                        artistaFechas.add(tempArtista);
+                                    }
+                                } catch (NoEsta e) {
+                                }
 
-                        try {
-                            procesador.hash().get(nombreA).setContdor(procesador.hash().get(nombreA).getContdor() + 1);
-                        } catch (NoEsta e) {
+                                try {
+                                    procesador.hash().get(nombreA).setContdor(procesador.hash().get(nombreA).getContdor() + 1);
+                                } catch (NoEsta e) {
+                                }
+                            }
                         }
                     }
                 }
+                tempfecha = tempfecha.plus(1, ChronoUnit.DAYS);
             }
-            tempfecha = tempfecha.plus(1, java.time.temporal.ChronoUnit.DAYS);
+            for (int i = 0; i < artistaFechas.size(); i++) {
+                Artista ArtistaTemp = artistaFechas.get(i);
+                heapArtistas.insert(ArtistaTemp, ArtistaTemp);
+            }
+            System.out.println("------------------------------------------------------");
+            for (int j = 0; j < 7; j++) {
+                Artista tempArtist = heapArtistas.delete();
+                System.out.println("TOP " + (j + 1) + " " + tempArtist.getNombre());
+            }
+            for (int i = 0; i < artistaFechas.size(); i++) {
+                Artista ArtistaTemp = artistaFechas.get(i);
+                ArtistaTemp.setContdor(0);
+            }
+            System.out.println("------------------------------------------------------");
         }
-        for(int i = 0; i<artistaFechas.size(); i++) {
-            Artista ArtistaTemp = artistaFechas.get(i);
-            heapArtistas.insert(ArtistaTemp, ArtistaTemp);
-        }
-        System.out.println("------------------------------------------------------");
-        for(int j = 0; j<7; j++){
-            Artista tempArtist = heapArtistas.delete();
-            System.out.println("TOP "+ (j+1)+" "+tempArtist.getNombre());
-        }
-        for(int i = 0; i<artistaFechas.size(); i++) {
-            Artista ArtistaTemp = artistaFechas.get(i);
-            ArtistaTemp.setContdor(0);
-        }
-        System.out.println("------------------------------------------------------");
     }
 
     public Object Top_5_canciones() {
@@ -130,6 +164,7 @@ public class Sistema {
     }
 
     public void cant_art() {
+        LocalDate localDate = null;
         System.out.println("------------------------------------------------------");
         Scanner scanner = new Scanner(System.in);
         System.out.println("Seleccione un artista");
@@ -137,25 +172,35 @@ public class Sistema {
         if (procesador.hash().contains(nombreA.toLowerCase())) {
             try {
                 Artista tempA = procesador.hash().get(nombreA.toLowerCase());
-                Scanner scanner1 = new Scanner(System.in);
-                System.out.println("Seleccione la fehca de inicio con el formato dd/mm/yyyy");
-                String fecha1 = scanner1.nextLine();
                 DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
-                LocalDate localDate = LocalDate.parse(fecha1, formatter);
+                while (localDate == null) {
+                    Scanner scanner1 = new Scanner(System.in);
+                    System.out.println("Seleccione la fehca con el formato dd/mm/yyyy");
+                    String fecha1 = scanner1.nextLine();
+                    try {
+                        localDate = LocalDate.parse(fecha1, formatter);
+                    } catch (Exception e) {
+                        System.out.println("La fecha no es compatible con el formato");
+                    }
+                }
                 ChronoLocalDate fecha1Formato = localDate;
-                Scanner scanner2 = new Scanner(System.in);
-                System.out.println("Seleccione la fehca de fin con el formato dd/mm/yyyy");
-                String fecha2 = scanner2.nextLine();
-                LocalDate localDate2 = LocalDate.parse(fecha2, formatter);
-                ChronoLocalDate fecha2Formato = localDate2;
-                ChronoLocalDate tempfecha = fecha1Formato;
-                while (!tempfecha.equals(fecha2Formato)) {
-                    tempfecha = tempfecha.plus(1, java.time.temporal.ChronoUnit.DAYS);
-                    MyList<Top50> recorrer = arbolBusqueda.find(tempfecha).Values();
+                LocalDate localdatepri = LocalDate.parse("18/10/2023", formatter);
+                ChronoLocalDate primeraFecha = localdatepri;
+                LocalDate localdateult = LocalDate.parse("14/05/2024", formatter);
+                ChronoLocalDate ultimaFecha = localdateult;
+                if (fecha1Formato.isBefore(primeraFecha)) {
+                    System.out.println("No tenemos registro de una fecha tan lejana");
+                }
+                if (fecha1Formato.isAfter(ultimaFecha)) {
+                    System.out.println("No tenemos registro de una fecha tan cercana");
+                }
+                if (arbolBusqueda.find(fecha1Formato) != null) {
+                    MyList<Top50> recorrer = arbolBusqueda.find(fecha1Formato).Values();
                     for (int i = 0; i < recorrer.size(); i++) {
                         Top50 tempTop = recorrer.get(i);
                         MyHeap<Cancion, Integer> clonTop = tempTop.getTop().clonar();
-                        for (int j = 0; j < 50; j++) {
+                        int tam = clonTop.size();
+                        for (int j = 0; j <tam; j++) {
                             Cancion canciontemp = clonTop.delete();
                             if (canciontemp.getArtist().contains(nombreA)) {
                                 tempA.setContdor(tempA.getContdor() + 1);
@@ -164,32 +209,48 @@ public class Sistema {
                         }
                     }
                 }
-                System.out.println("El artista aparece "+ tempA.getContdor()+ " veces");
+                System.out.println("El artista aparece " + tempA.getContdor() + " veces");
                 tempA.setContdor(0);
             } catch (NoEsta e) {
             }
-        }else {
+        } else {
             System.out.println("Ese artista no esta");
         }
+
     }
 
 
-
-    public int cant_canc_tempo (){
+    public int cant_canc_tempo () {
         System.out.println("------------------------------------------------------");
         MyList<String> cancionesYaContadas = new MyLinkedListImpl<>();
         int cantcan = 0;
-        Scanner scanner1 = new Scanner(System.in);
-        System.out.println("Seleccione la fehca de inicio con el formato dd/mm/yyyy");
-        String fecha1 = scanner1.nextLine();
+        LocalDate localDate = null;
+        LocalDate localDate2 = null;
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
-        LocalDate localDate = LocalDate.parse(fecha1, formatter);
+        while (localDate == null) {
+            Scanner scanner1 = new Scanner(System.in);
+            System.out.println("Seleccione la fehca de inicio con el formato dd/mm/yyyy");
+            String fecha1 = scanner1.nextLine();
+            try {
+                localDate = LocalDate.parse(fecha1, formatter);
+            } catch (Exception e) {
+                System.out.println("La fecha no es compatible con el formato");
+            }
+
+        }
         ChronoLocalDate fecha1Formato = localDate;
-        Scanner scanner2 = new Scanner(System.in);
-        System.out.println("Seleccione la fehca de fin con el formato dd/mm/yyyy");
-        String fecha2 = scanner2.nextLine();
-        LocalDate localDate2 = LocalDate.parse(fecha2, formatter);
+        while (localDate2 == null) {
+            Scanner scanner2 = new Scanner(System.in);
+            System.out.println("Seleccione la fehca de fin con el formato dd/mm/yyyy");
+            String fecha2 = scanner2.nextLine();
+            try {
+                localDate2 = LocalDate.parse(fecha2, formatter);
+            } catch (Exception e) {
+                System.out.println("La fecha no es compatible con el formato");
+            }
+        }
         ChronoLocalDate fecha2Formato = localDate2;
+        ChronoLocalDate tempfecha = fecha1Formato;
         Scanner scanner3 = new Scanner(System.in);
         System.out.println("Seleccione el menor del rengo del tempo");
         String tempS1 = scanner3.nextLine();
@@ -198,29 +259,46 @@ public class Sistema {
         System.out.println("Seleccione el mayor del rengo del tempo");
         String tempS2 = scanner4.nextLine();
         Float temp2 = Float.parseFloat(tempS2);
-        ChronoLocalDate tempfecha = fecha1Formato;
-        while (!tempfecha.equals(fecha2Formato)){
-            MyList<Top50> recorrer = arbolBusqueda.find(tempfecha).Values();
-            for(int i = 0; i<recorrer.size(); i++){
-                Top50 tempTop = recorrer.get(i);
-                MyHeap<Cancion, Integer> clonTop = tempTop.getTop().clonar();
-                for(int j = 0; j<50; j++){
-                    Cancion canciontemp = clonTop.delete();
-                    Float tempo = canciontemp.getTempo();
-                    if(tempo>temp1 && tempo<temp2){
-                        if(!cancionesYaContadas.contains(canciontemp.getName())){
-                            cantcan++;
-                            cancionesYaContadas.add(canciontemp.getName());
-                        }
+        LocalDate localdatepri = LocalDate.parse("18/10/2023", formatter);
+        ChronoLocalDate primeraFecha = localdatepri;
+        LocalDate localdateult = LocalDate.parse("14/05/2024", formatter);
+        ChronoLocalDate ultimaFecha = localdateult;
+        if (fecha1Formato.isAfter(fecha2Formato)) {
+            System.out.println("No tiene sentido el orden");
+        }
+        if (fecha1Formato.isBefore(primeraFecha)) {
+            System.out.println("No tenemos registro de una fecha tan lejana");
+        }
+        if (fecha2Formato.isAfter(ultimaFecha)) {
+            System.out.println("No tenemos registro de una fecha tan cercana");
+        } else {
+            while (!tempfecha.equals(fecha2Formato)) {
+                if (arbolBusqueda.find(tempfecha) != null) {
+                    MyList<Top50> recorrer = arbolBusqueda.find(tempfecha).Values();
+                    for (int i = 0; i < recorrer.size(); i++) {
+                        Top50 tempTop = recorrer.get(i);
+                        MyHeap<Cancion, Integer> clonTop = tempTop.getTop().clonar();
+                        int tam = clonTop.size();
+                        for (int j = 0; j < tam; j++) {
+                            Cancion canciontemp = clonTop.delete();
+                            Float tempo = canciontemp.getTempo();
+                            if (tempo > temp1 && tempo < temp2) {
+                                if (!cancionesYaContadas.contains(canciontemp.getName())) {
+                                    cantcan++;
+                                    cancionesYaContadas.add(canciontemp.getName());
+                                }
 
+                            }
+                        }
                     }
                 }
+                tempfecha = tempfecha.plus(1, ChronoUnit.DAYS);
             }
-            tempfecha = tempfecha.plus(1, java.time.temporal.ChronoUnit.DAYS);
+            System.out.println("------------------------------------------------------");
+            System.out.println("Hay " + cantcan + " En ese rango de fechas con ese rango de tempo");
+            System.out.println("------------------------------------------------------");
+            return cantcan;
         }
-        System.out.println("------------------------------------------------------");
-        System.out.println("Hay "+ cantcan+" En ese rango de fechas con ese rango de tempo");
-        System.out.println("------------------------------------------------------");
-        return cantcan;
+        return 0;
     }
 }
